adjusts
Represents the SCM configuration in a multibranch project build. Use checkout scm to check out sources matching Jenkinsfile. 
You may also use this in a standalone project configured with Pipeline script from SCM, though in that case the checkout will just be of the latest revision in the branch, possibly newer than the revision from which the Pipeline script was loaded.